Replace `EntryMetadataId` with `EntryHandle`, refactor all impacted classes, and switch to Murmur3_128 for handle generation. Add `EntryHandleGenerator` and update tests for compatibility.

+++ kronotop/src/main/java/com/kronotop/bucket/pipeline/DocumentLocationSink.java
@@ -16,7 +16,7 @@
 
 package com.kronotop.bucket.pipeline;
 
-public final class DocumentLocationSink extends AbstractDataSink<Integer, DocumentLocation> implements DataSink {
+public final class DocumentLocationSink extends AbstractDataSink<Long, DocumentLocation> implements DataSink {
     public DocumentLocationSink(int parentNodeId) {
         super(parentNodeId);
     }
